requirements.txt
psycopg2_binary==2.9.9
python-dotenv==1.0.1
Requests==2.31.0
